Phase 20 Batch 2: Extract consolidatable guards from GCI0038, GCI0041, GCI0022
- Added 5 new guard methods to GuardPatterns class
  - IsMigrationOrSeedFile(): detects migration/seed data files (GCI0022)
  - IsUiEventHandler(): detects UI/XAML context files (GCI0022)
  - IsDocumentationFile(): detects doc/snippet/sample files (GCI0041)
  - IsInsideStaticConstructor(): detects static constructors (GCI0022)
  - Enhanced IsEventHandler(): now includes EventArgs detection for better event handler recognition

- Refactored 3 rules to use consolidated guards
  - GCI0038: Replaced inline IsTestFile() with WellKnownPatterns.IsTestFile()
  - GCI0041: Replaced IsTestFile() and removed documentation skip logic in favor of GuardPatterns
  - GCI0022: Replaced IsMigrationOrSeedFile(), IsInsideStaticConstructor(), IsUiEventHandler() with GuardPatterns

- Fixed HasAccessModifier() to handle attributes
  - Now skips leading attributes like [Obsolete] before checking for access modifiers
  - Handles methods with multiple attributes: [Attr1][Attr2] public void Method()

- LOC removed from rule implementations: ~130 lines
- LOC added to GuardPatterns: ~160 lines
- Net consolidation: +30 lines but enables reuse across all rules
- Tests: 1,275/1,275 passing ✅
