+++ src/GauntletCI.Core/Rules/Implementations/GCI0022_IdempotencyRetrySafety.cs
@@ -70,7 +70,7 @@ public class GCI0022_IdempotencyRetrySafety : RuleBase
     private void CheckRawInsertWithoutUpsert(DiffFile file, List<Finding> findings)
     {
         // Skip migration and seed data files - they use raw INSERT intentionally
-        if (IsMigrationOrSeedFile(file.NewPath))
+        if (WellKnownPatterns.GuardPatterns.IsMigrationOrSeedFile(file.NewPath))
             return;
 
         foreach (var line in file.AddedLines)
@@ -111,10 +111,10 @@ public class GCI0022_IdempotencyRetrySafety : RuleBase
                 !contentLower.Contains("listener") && !contentLower.Contains("callback")) continue;
 
             // Exempt += inside a static constructor (runs exactly once -- inherently idempotent)
-            if (IsInsideStaticConstructor(allLines, i)) continue;
+            if (WellKnownPatterns.GuardPatterns.IsInsideStaticConstructor(allLines, i)) continue;
 
             // Exempt if in UI/XAML context (WPF, WinUI events are often attached once per control lifecycle)
-            if (IsUiEventHandler(file.NewPath)) continue;
+            if (WellKnownPatterns.GuardPatterns.IsUiEventHandler(file.NewPath)) continue;
 
             // Look for deduplication guard nearby (unsubscribe or bool guard)
             int start = Math.Max(0, i - 5);
@@ -139,81 +139,5 @@ public class GCI0022_IdempotencyRetrySafety : RuleBase
                     line: line));
             }
         }
-    }
-
-    private static bool IsMigrationOrSeedFile(string filePath)
-    {
-        // Migration files: EF Core migrations directory
-        if (filePath.Contains("Migrations/", StringComparison.OrdinalIgnoreCase) ||
-            filePath.Contains("\\Migrations\\", StringComparison.OrdinalIgnoreCase))
-            return true;
-
-        // Migration SQL scripts
-        if (filePath.EndsWith(".sql", StringComparison.OrdinalIgnoreCase) &&
-            (filePath.Contains("Migration", StringComparison.OrdinalIgnoreCase) ||
-             filePath.Contains("Seed", StringComparison.OrdinalIgnoreCase) ||
-             filePath.Contains("Setup", StringComparison.OrdinalIgnoreCase)))
-            return true;
-
-        // EF seed configurations (ModelBuilder.Entity.HasData)
-        if (filePath.Contains("SeedData", StringComparison.OrdinalIgnoreCase) ||
-            filePath.Contains("DataSeeding", StringComparison.OrdinalIgnoreCase))
-            return true;
-
-        return false;
-    }
-
-    /// <summary>
-    /// Returns true when the line at <paramref name="idx"/> appears to be inside a static constructor,
-    /// which is inherently idempotent (runs exactly once per AppDomain lifetime).
-    /// Detects the pattern "static ClassName()" or "static()" within the preceding 20 lines.
-    /// </summary>
-    private static bool IsInsideStaticConstructor(List<DiffLine> allLines, int idx)
-    {
-        int searchStart = Math.Max(0, idx - 20);
-        for (int j = idx - 1; j >= searchStart; j--)
-        {
-            var trimmed = allLines[j].Content.Trim();
-            if (!trimmed.StartsWith("static ", StringComparison.Ordinal)) continue;
-
-            var afterStatic = trimmed["static ".Length..].TrimStart();
-            // If the next token is a C# keyword that can be a return type, it's a method not a constructor
-            if (afterStatic.StartsWith("void ", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("Task", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("bool ", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("int ", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("string ", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("async ", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("readonly ", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("class ", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("IEnumerable", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("List<", StringComparison.Ordinal) ||
-                afterStatic.StartsWith("Dictionary<", StringComparison.Ordinal))
-                continue;
-
-            // The remainder should look like "TypeName()" - contains "()"
-            if (afterStatic.Contains("()")) return true;
-        }
-        return false;
-    }
-
-    /// <summary>
-    /// Returns true if the file is a UI/XAML context (WPF, WinUI, Blazor)
-    /// where event handler registration patterns are often benign.
-    /// </summary>
-    private static bool IsUiEventHandler(string filePath)
-    {
-        var lower = filePath.ToLowerInvariant();
-        // XAML code-behind files
-        if (lower.EndsWith(".xaml.cs", StringComparison.OrdinalIgnoreCase)) return true;
-        // Blazor component files
-        if (lower.EndsWith(".razor.cs", StringComparison.OrdinalIgnoreCase)) return true;
-        // Common UI namespaces
-        if (lower.Contains("\\ui\\") || lower.Contains("/ui/") ||
-            lower.Contains("\\components\\") || lower.Contains("/components/") ||
-            lower.Contains("\\views\\") || lower.Contains("/views/") ||
-            lower.Contains("\\pages\\") || lower.Contains("/pages/")) return true;
-
-        return false;
     }
 }
